backend/requirements.txt
autopep8==2.0.2
bidict==0.22.1
click==8.1.3
Flask==2.2.3
Flask-SocketIO==5.3.3
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.2
gunicorn==20.1.0
importlib-metadata==6.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pycodestyle==2.10.0
python-engineio==4.4.0
python-socketio==5.8.0
pyzmq==25.0.2
tomli==2.0.1
Werkzeug==2.2.3
zipp==3.15.0
zope.event==4.6
zope.interface==6.0
